empty() does not clear the array and checks whether it is empty.

+++ ogsr_engine/xrGame/BoneProtections.cpp
@@ -45,7 +45,7 @@ void SBoneProtections::reload(const shared_str& bone_sect, CKinematics* kinemati
 		string256 buffer;
 		float Koeff = (float)atof( _GetItem(*(*i).second, 0, buffer) );
 		float Armour = (float)atof( _GetItem(*(*i).second, 1, buffer) );
-		BOOL BonePassBullet = (BOOL) (atoi( _GetItem(*(*i).second, 2, buffer) )>0.5f);
+		BOOL BonePassBullet = (BOOL) (atof( _GetItem(*(*i).second, 2, buffer) )>0.5f);
 		
 		BoneProtection	BP;
 		BP.koeff = Koeff;
